move avatar list item to features users
