+++ web-admin/src/features/organizations/users/AvatarListItem.svelte
