+++ web-admin/src/features/projects/user-invite/UserInviteGroupAvatar.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { createAdminServiceListUsergroupMemberUsers } from "@rilldata/web-admin/client";
-  import AvatarListItem from "@rilldata/web-common/components/avatar/AvatarListItem.svelte";
+  import AvatarListItem from "../../organizations/users/AvatarListItem.svelte";
 
   export let organization: string;
   export let usergroup: string;
